Returned the properly typed last inserted rowid
For proper type safety `storage_t<>::insert()` needs to return an `sqlite3_int64` instead of an `int`.
This is the best way to correct the problem. Usually, compilers just accept narrowing type conversions, the nicest thing that can happen is a warning message, the worst is an error depending on compiler flags. The programmer has to handle the rest and it is the best thing to do anyway.

+++ dev/storage.h
@@ -941,11 +941,11 @@ namespace sqlite_orm {
             }
 
             template<class O, class... Cols>
-            int insert(const O& o, columns_t<Cols...> cols) {
+            int64 insert(const O& o, columns_t<Cols...> cols) {
                 static_assert(cols.count > 0, "Use insert or replace with 1 argument instead");
                 this->assert_mapped_type<O>();
                 auto statement = this->prepare(sqlite_orm::insert(std::ref(o), std::move(cols)));
-                return int(this->execute(statement));
+                return this->execute(statement);
             }
 
             /**
@@ -961,15 +961,13 @@ namespace sqlite_orm {
              *  
              *  @return The ID of the last inserted record for a table with rowid, otherwise a meaningless value.
              *          Attention: `sqlite3_last_insert_rowid()` is used to retrieve the last inserted ID, therefore the ID is only useful in single-threaded contexts.
-             *          Attention: While SQLite returns a 64-bit integer as rowid, this function returns an `int` that most likely has less precision.
-             *                     If you need the full 64-bit rowid, use `storage_t<>::execute()` instead.
              */
             template<class O>
-            int insert(const O& o) {
+            int64 insert(const O& o) {
                 this->assert_mapped_type<O>();
                 this->assert_insertable_type<O>();
                 auto statement = this->prepare(sqlite_orm::insert(std::ref(o)));
-                return int(this->execute(statement));
+                return this->execute(statement);
             }
 
             /**
